COMMERCE-8636 account-service: moves read/write logic to keep the CurrentAccountEntryManagerImpl out of the permission stack
