+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/manager/CurrentAccountEntryManagerImpl.java
@@ -15,7 +15,6 @@
 package com.liferay.account.internal.manager;
 
 import com.liferay.account.constants.AccountConstants;
-import com.liferay.account.constants.AccountWebKeys;
 import com.liferay.account.exception.AccountEntryTypeException;
 import com.liferay.account.manager.CurrentAccountEntryManager;
 import com.liferay.account.model.AccountEntry;
@@ -26,25 +25,17 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.UserConstants;
-import com.liferay.portal.kernel.portlet.PortalPreferences;
-import com.liferay.portal.kernel.portlet.PortletPreferencesFactory;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.security.permission.PermissionCheckerFactory;
 import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.service.PortalPreferencesLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
-import com.liferay.portal.kernel.servlet.PortalSessionThreadLocal;
 import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
 import java.util.List;
 import java.util.Objects;
-
-import javax.servlet.http.HttpSession;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -78,7 +69,9 @@ public class CurrentAccountEntryManagerImpl
 		PermissionChecker permissionChecker = _permissionCheckerFactory.create(
 			user);
 
-		AccountEntry accountEntry = _getAccountEntryFromHttpSession(groupId);
+		AccountEntry accountEntry =
+			_currentAccountEntryManagerStore.getAccountEntryFromHttpSession(
+				groupId);
 
 		String[] allowedTypes = _getAllowedTypes(groupId);
 
@@ -86,10 +79,13 @@ public class CurrentAccountEntryManagerImpl
 			return accountEntry;
 		}
 
-		accountEntry = _getAccountEntryFromPortalPreferences(groupId, userId);
+		accountEntry =
+			_currentAccountEntryManagerStore.
+				getAccountEntryFromPortalPreferences(groupId, userId);
 
 		if (_isValid(accountEntry, allowedTypes, permissionChecker)) {
-			_saveInHttpSession(accountEntry.getAccountEntryId(), groupId);
+			_currentAccountEntryManagerStore.saveInHttpSession(
+				accountEntry.getAccountEntryId(), groupId);
 
 			return accountEntry;
 		}
@@ -126,45 +122,14 @@ public class CurrentAccountEntryManagerImpl
 						"type: " + accountEntry.getType());
 			}
 
-			_saveInHttpSession(accountEntryId, groupId);
-
-			_saveInPortalPreferences(accountEntryId, groupId, userId);
+			_currentAccountEntryManagerStore.setCurrentAccountEntry(
+				accountEntryId, groupId, userId);
 		}
 		catch (PortalException portalException) {
 			if (_log.isWarnEnabled()) {
 				_log.warn(portalException);
 			}
 		}
-	}
-
-	private AccountEntry _getAccountEntryFromHttpSession(long groupId) {
-		HttpSession httpSession = PortalSessionThreadLocal.getHttpSession();
-
-		if (httpSession == null) {
-			return null;
-		}
-
-		long currentAccountEntryId = GetterUtil.getLong(
-			httpSession.getAttribute(_getKey(groupId)));
-
-		return _accountEntryLocalService.fetchAccountEntry(
-			currentAccountEntryId);
-	}
-
-	private AccountEntry _getAccountEntryFromPortalPreferences(
-		long groupId, long userId) {
-
-		PortalPreferences portalPreferences = _getPortalPreferences(userId);
-
-		long accountEntryId = GetterUtil.getLong(
-			portalPreferences.getValue(
-				AccountEntry.class.getName(), _getKey(groupId)));
-
-		if (accountEntryId > 0) {
-			return _accountEntryLocalService.fetchAccountEntry(accountEntryId);
-		}
-
-		return null;
 	}
 
 	private String[] _getAllowedTypes(long groupId) {
@@ -185,14 +150,6 @@ public class CurrentAccountEntryManagerImpl
 		}
 
 		return null;
-	}
-
-	private String _getKey(long groupId) {
-		return AccountWebKeys.CURRENT_ACCOUNT_ENTRY_ID + groupId;
-	}
-
-	private PortalPreferences _getPortalPreferences(long userId) {
-		return _portletPreferencesFactory.getPortalPreferences(userId, true);
 	}
 
 	private boolean _isValid(
@@ -223,39 +180,6 @@ public class CurrentAccountEntryManagerImpl
 		return false;
 	}
 
-	private void _saveInHttpSession(long accountEntryId, long groupId) {
-		HttpSession httpSession = PortalSessionThreadLocal.getHttpSession();
-
-		if (httpSession == null) {
-			return;
-		}
-
-		httpSession.setAttribute(_getKey(groupId), accountEntryId);
-	}
-
-	private void _saveInPortalPreferences(
-		long accountEntryId, long groupId, long userId) {
-
-		PortalPreferences portalPreferences = _getPortalPreferences(userId);
-
-		String key = _getKey(groupId);
-
-		long currentAccountEntryId = GetterUtil.getLong(
-			portalPreferences.getValue(
-				AccountEntry.class.getName(), key,
-				String.valueOf(AccountConstants.ACCOUNT_ENTRY_ID_GUEST)));
-
-		if (currentAccountEntryId == accountEntryId) {
-			return;
-		}
-
-		portalPreferences.setValue(
-			AccountEntry.class.getName(), key, String.valueOf(accountEntryId));
-
-		_portalPreferencesLocalService.updatePreferences(
-			userId, PortletKeys.PREFS_OWNER_TYPE_USER, portalPreferences);
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		CurrentAccountEntryManagerImpl.class);
 
@@ -272,13 +196,10 @@ public class CurrentAccountEntryManagerImpl
 		_accountEntryModelResourcePermission;
 
 	@Reference
+	private CurrentAccountEntryManagerStore _currentAccountEntryManagerStore;
+
+	@Reference
 	private PermissionCheckerFactory _permissionCheckerFactory;
-
-	@Reference
-	private PortalPreferencesLocalService _portalPreferencesLocalService;
-
-	@Reference
-	private PortletPreferencesFactory _portletPreferencesFactory;
 
 	@Reference
 	private UserLocalService _userLocalService;
