+++ modules/apps/account/account-service/src/main/java/com/liferay/account/internal/manager/CurrentAccountEntryManagerStore.java
@@ -1,0 +1,140 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.account.internal.manager;
+
+import com.liferay.account.constants.AccountConstants;
+import com.liferay.account.constants.AccountWebKeys;
+import com.liferay.account.model.AccountEntry;
+import com.liferay.account.service.AccountEntryLocalService;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.portlet.PortalPreferences;
+import com.liferay.portal.kernel.portlet.PortletPreferencesFactory;
+import com.liferay.portal.kernel.service.PortalPreferencesLocalService;
+import com.liferay.portal.kernel.servlet.PortalSessionThreadLocal;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.PortletKeys;
+
+import javax.servlet.http.HttpSession;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Drew Brokke
+ */
+@Component(service = CurrentAccountEntryManagerStore.class)
+public class CurrentAccountEntryManagerStore {
+
+	public AccountEntry getAccountEntryFromHttpSession(long groupId) {
+		HttpSession httpSession = PortalSessionThreadLocal.getHttpSession();
+
+		if (httpSession == null) {
+			return null;
+		}
+
+		long currentAccountEntryId = GetterUtil.getLong(
+			httpSession.getAttribute(_getKey(groupId)));
+
+		return _accountEntryLocalService.fetchAccountEntry(
+			currentAccountEntryId);
+	}
+
+	public AccountEntry getAccountEntryFromPortalPreferences(
+		long groupId, long userId) {
+
+		PortalPreferences portalPreferences = _getPortalPreferences(userId);
+
+		long accountEntryId = GetterUtil.getLong(
+			portalPreferences.getValue(
+				AccountEntry.class.getName(), _getKey(groupId)));
+
+		if (accountEntryId > 0) {
+			return _accountEntryLocalService.fetchAccountEntry(accountEntryId);
+		}
+
+		return null;
+	}
+
+	public AccountEntry getCurrentAccountEntry(long groupId, long userId)
+		throws PortalException {
+
+		AccountEntry accountEntry = getAccountEntryFromHttpSession(groupId);
+
+		if (accountEntry == null) {
+			accountEntry = getAccountEntryFromPortalPreferences(
+				groupId, userId);
+		}
+
+		return accountEntry;
+	}
+
+	public void saveInHttpSession(long accountEntryId, long groupId) {
+		HttpSession httpSession = PortalSessionThreadLocal.getHttpSession();
+
+		if (httpSession == null) {
+			return;
+		}
+
+		httpSession.setAttribute(_getKey(groupId), accountEntryId);
+	}
+
+	public void saveInPortalPreferences(
+		long accountEntryId, long groupId, long userId) {
+
+		PortalPreferences portalPreferences = _getPortalPreferences(userId);
+
+		String key = _getKey(groupId);
+
+		long currentAccountEntryId = GetterUtil.getLong(
+			portalPreferences.getValue(
+				AccountEntry.class.getName(), key,
+				String.valueOf(AccountConstants.ACCOUNT_ENTRY_ID_GUEST)));
+
+		if (currentAccountEntryId == accountEntryId) {
+			return;
+		}
+
+		portalPreferences.setValue(
+			AccountEntry.class.getName(), key, String.valueOf(accountEntryId));
+
+		_portalPreferencesLocalService.updatePreferences(
+			userId, PortletKeys.PREFS_OWNER_TYPE_USER, portalPreferences);
+	}
+
+	public void setCurrentAccountEntry(
+		long accountEntryId, long groupId, long userId) {
+
+		saveInHttpSession(accountEntryId, groupId);
+		saveInPortalPreferences(accountEntryId, groupId, userId);
+	}
+
+	private String _getKey(long groupId) {
+		return AccountWebKeys.CURRENT_ACCOUNT_ENTRY_ID + groupId;
+	}
+
+	private PortalPreferences _getPortalPreferences(long userId) {
+		return _portletPreferencesFactory.getPortalPreferences(userId, true);
+	}
+
+	@Reference
+	private AccountEntryLocalService _accountEntryLocalService;
+
+	@Reference
+	private PortalPreferencesLocalService _portalPreferencesLocalService;
+
+	@Reference
+	private PortletPreferencesFactory _portletPreferencesFactory;
+
+}
